+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # Change log
 
-## v0.3.2 3.11.2018
+## v0.5.1 4.11.2018
+
+**Added:**
+- Dictionaries for user roles
+- AccountType to customer model
+- Account type in registratiion process
+
+**Fixed:**
+- registration validation (incorrect role type)
+
+## v0.5.0 3.11.2018
 
 **Added:**
 - Model for Resources
 - Dictionary items for Language
 - Endpoint to get list of static resources, list of static resources in selected language, list of resource in every languages, single resource in selected language
 
-## v0.3.1 31.10.2018
+## v0.4.0 31.10.2018
 
 **Added:**
 - Model for Dictionary and DictionaryItem